+++ lib/player/simple/simple_player.dart
@@ -53,7 +53,6 @@ class SimplePlayer extends Player {
       }
     }
     lastDirection = Direction.up;
-
     super.moveUp(speed, onCollision: onCollision);
   }
 
@@ -97,7 +96,8 @@ class SimplePlayer extends Player {
     }
     lastDirection = Direction.upLeft;
     lastDirectionHorizontal = Direction.left;
-    super.moveUpLeft(speedX, speedY, onCollision: onCollision);
+    super.moveUp(speedY, onCollision: onCollision);
+    super.moveLeft(speedX, onCollision: onCollision);
   }
 
   @override
@@ -109,7 +109,8 @@ class SimplePlayer extends Player {
     }
     lastDirection = Direction.upRight;
     lastDirectionHorizontal = Direction.right;
-    super.moveUpRight(speedX, speedY, onCollision: onCollision);
+    super.moveUp(speedY, onCollision: onCollision);
+    super.moveRight(speedX, onCollision: onCollision);
   }
 
   @override
@@ -122,7 +123,8 @@ class SimplePlayer extends Player {
     }
     lastDirection = Direction.downRight;
     lastDirectionHorizontal = Direction.right;
-    super.moveDownRight(speedX, speedY, onCollision: onCollision);
+    super.moveDown(speedY, onCollision: onCollision);
+    super.moveRight(speedX, onCollision: onCollision);
   }
 
   @override
@@ -134,7 +136,8 @@ class SimplePlayer extends Player {
     }
     lastDirection = Direction.downLeft;
     lastDirectionHorizontal = Direction.left;
-    super.moveDownLeft(speedX, speedY, onCollision: onCollision);
+    super.moveDown(speedY, onCollision: onCollision);
+    super.moveLeft(speedX, onCollision: onCollision);
   }
 
   @override